requirements.txt
langchain==0.2.17
langchain-community==0.2.17
google-generativeai
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic
langchain-google-genai
dateparser
flask